refactor(spec): Split the spec file in to two
The original spec file was performing two types of tests, i've split it in to a file checking if the
Gulp file is adding tasks as expected. The original file now only checks the output of running the
task
